Update google-java-format to 1.2-SNAPSHOT.
This brings support for Javadoc formatting and other bug fixes. We use
a snapshot because the --skip-sorting-imports flag was just recently
added, but is important for the formatter to be useful in the AOSP
codebase as the AOSP import style differs from the google style.
(Otherwise, apart from indenting, which already can be controlled with
the --aosp flag, the styles are effectively consistent).

For now, keep the old 1.0 .jar around, as there are some scripts which
depend on the exact path to it. It will be removed once those scripts
are updated or deleted.

Bug: 31552314
Test: Verified with local repository
Change-Id: I9e93194f785766fcb0d73a2586201c8a0aae1f1d