Update AAPT2 JNI prebuilts
Branch git_master-release has failing targets and therefore we cannot
grab the prebuilts from there. The only other release branch that
contains the same targets that are not failing is git_oc-release. It
contains the fix we need, but it does not contain prebuilts for
Windows64.
We will need to update the prebuilts as soon as git_master-release is
green again.

Test: LinkTest, CompileTest
BUG=38173018

Change-Id: I2d90543f57b49299caf5e228ce4ed64fef3d0440