Add support libraries to prebuilts repo
This CL adds the support libraries required by TemplateTest
to the prebuilts repo. These files were previously in the SDK,
however we now host them in the google maven repo and as such
these files need to now be here as they will no longer be added
to prebuilts/studio/sdk.

The contents of this CL was generated with the following command:

bazel run //tools/base/bazel:add_dependency \
  com.android.support:appcompat-v7:aar:26.0.0-beta1 \
  com.android.support:design:aar:26.0.0-beta1 \
  com.android.support:leanback-v17:aar:26.0.0-beta1 \
  com.android.support:support-v13:aar:26.0.0-beta1 \
  com.android.support:percent:aar:26.0.0-beta1 \
  com.android.support:palette-v7:aar:26.0.0-beta1

Test: TemplateTest

Change-Id: I083382a24aade5556a799c4f92ef78f9ba1105b7