Store Kotlin plugin unpacked
The benefit of having it unpacked is that we can use it in local
Studio / unit / UI test runs via -Dplugin.path (see PluginManagerCore).

Currently it gets unzipped for every build anyway. For local Studio
runs we have to manually install it into idea.config.path, and
depending on how the config folder gets cleaned or moved, the Kotlin
install may not survive.

Test: N/A
Change-Id: I722d6b50515c1c2fd2a7cad974a8ae337eadbc6d