Add LICENSE and NOTICE to trove4j GroupId dir
The NOTICE file for trove4j is currently in the artifact directory, not
the groupId directory.  This is causing gradle build to fail as it does
not properly find this License and :base:gradle:licenseReport reports a
error.  This checkin copies this LICENSE up one directory, and also adds
the NOTICE file (a direct copy of LICENSE) to match other kotlin based
libraries in the the same dir structure.

Change-Id: I4f88afe9769d38ca13b09022f1e505b16c46dc9c
(cherry picked from commit 71e88e9141e3e24df23785981fd3923172ab7dc5)