Update Android support library metadata
The metadata files became outdated for the Android support libraries in
the prebuilts Maven repository. This can cause UI test projects to use
outdated libraries when running tests, which means the tests will not
catch compatibility issues with newer libraries.

In order for our tests to test against the newer libraries, the metadata
files need to be up-to-date.

Change-Id: Ie3cf73a93c98808e46a2c490ee40c7f0f96a79d2
Fixes: 74400675
Test: all existing tests. Ran sanity tests locally