Crashlytics: Add Crashlytics plugin 1.25.4 to prebuilts.
In the other CL of the same topic, we enforce minimum version 1.25.4 for
the Crashlytics plugin, so adding that version here for testing.

Note that we still keep the old version (1.22.1) in prebuilts as it is
used in a regression test.

Bug: 79997489
Test: PluginVersionCheckTest (See the other CL)
Change-Id: I7635cd618030640578eb32b57d4ea56795c63f13
(cherry picked from commit ee53849a6f6103a29bf99b08ae59ba7bc397c7ef)