Add protoc3.6 prebuilts to be used by trusty app.
Trusty app uses a different build system than Android, thus it is
very hard to build protoc in trusty repo. Trusty applications need
to use nanopb to serialize data and write to secure storage (RPMB).
nanopb depends on protoc to translate the .proto file to .h and .c
file, as a result, we need a protoc prebuilts. The current protoc
is version 2.5 and is outdated. It does not support some of the new
key words such as reserve which the nanopb descriptor file requires.

Test: None
Bug: b/110897607
Change-Id: I72715fdb46b6a49c028ed3218a604bfe832ca4ae