Update simpleperf windows binaries to include crucial fix
Simpleperf windows binaries released with NDK 19 contains a bug caused by
path separator and newline characters. The impact of that issue is high
because simpleperf report-sample will fail silently, generating a non-empty
trace file that is malformed and cannot be parsed by working simpleperf
trace parsers.

That was fixed by http://aosp/806215 and will be included in the next NDK
release, but we need simpleperf report-sample working on Studio before that.
We should update our windows binaries because having the risk of facing some
instability (as we're not getting simpleperf from a stable release) is
better than the feature not working at all.

These simpleperf binaries were downloaded from AOSP master (build 5134375)
artifacts.

Fixes: 119611493
Test: Manual. This was verified on a windows machine. Once the CL is merged,
      make sure the SIMPLEPERF_HOST flag is enabled and test native
      profiling on a windows machine.

Change-Id: I0752be470bc8c12dccc7e8994988260e8e180968