Strip unnecessary symbols for simpleperf on Windows
When simpleperf is updated in NDK, simpleperf.exe and
libsimpleperf_report.dll are stripped of unnecessary symbols. Because
simpleperf was updated directly from the AOSP SDK build, the stripping
was not done which caused the binaries to be much larger.
This replaces those binaries with stripped ones.
The simpleperf binaries before stripping were taken from build 5134375
of aosp-master.

Bug: 119611493
Test: manually tested that native profiling still works on a Windows
machine.

Change-Id: If13a57bfd6cbdfe929aa454af8a7d41ab9cd48b8