Upgrade to bazel version 0.20.0 with for Linux and patched Mac bazel version.
The linux binary is the latest public releases. The last couple mac public
releases had a bug that caused the macs to spawn too many threads which would
bring down the machine. That was fixed with this patch:
https://github.com/bazelbuild/bazel/commit/377670e89e2a89a6b656d1d5adf00ba48cabd08e
but it is not in 0.20.0 and may not even be in 0.21. Thus, the mac binary being
checked in here is the public version + that patch. I built checkout the source,
then did git checkout 0.20.0, and then applied the patch, and build the binary.
Windows is staying back on bazel version 0.19.2.

Bug: 118402489
Test: tools/base/bazel test //tools/...
Change-Id: I53a591d9eda3e852d108c3f8583708c13f681a07