Bump Perfetto version and expose official bazel rules.
Perfetto now can be built with bazel, but it has a number of external dependencies that are currently downloaded from http or git.
For reference:
https://googleplex-android.googlesource.com/platform/external/perfetto/+/refs/heads/master/bazel/deps.bzl

This CL pulls in the external dependencies required to build the TraceProcessor C++ library and protos.
These dependencies are referenced via bazel's http_archive rule, this CL drops in those archives (rather than creating a separate git folder for each dependency).

Test: manually with P device + manual trace loading in Studio
Bug: 142054929
Change-Id: I466fd7eb32a929f2dd60783c5a46c0d8077b7307