Patched bazel binary to modify worker-sandboxing behavior.
When a strategy is run with dynamic scheduling it runs two strategies
simultaneously, one remote and one local. Both of these strategies are
run within a symlinked sandbox to prevent the other strategy from
overwriting the outputs generated. When sandboxing workers on Windows,
the bazel implementation uses filesystem Junctions instead of true
symbolic links. This causes an unrecoverable failure when bazel tries
to re-use the worker sandbox for additional actions. Before each action
in the sandbox, bazel deletes all files and re-creates all necessary
inputs as symlinks. However since these symlinks are actually
Junctions on Windows, bazel ends up trying to delete the root file.

This modification to bazel stops the clearing of the worker sandbox
between actions. This is how the ordinary worker strategy behaves when
not sandboxed and should not lead to any unforseen side-effects. This
change is itself a (hopefully) temporary one until we can upstream a
bazel change which uses real symbolic linking on Windows. b/153345417

Dynamic scheduling improves the performance of actions which support
persistent workers. Using workers we've observed a 50-60% speed
improvement for Javac and kotlinc actions. This greatly reduces
the 90%ile latency when many chained dependencies need to be rebuilt.

Built with command:
bazel-1.1.0.exe build //src:bazel.exe ^
 -c opt --stamp --embed_label 1.1.0-adt-b147615404

Bug: 147615404
Test: Verified presubmit
Change-Id: I68088e95dbe445a0e7fd7f488596c3e1a0558611