Import the gRPC Netty Shaded library jar
Import the Netty Shaded library jar that is required for Transport Layer
Security (TLS) on non-Android platforms. The library prepackages
netty-tcnative on BoringSSL which makes it easier to use.

Note that although we compile the rest of gRPC Java from source we
import this particular jar since building it is non-trivial due to the
large number of external dependencies and missing build-system support.

The artifacts were imported using the below command:

${ANDROID_BUILD_TOP}/prebuilts/tools/import-maven-artifacts.sh \
  io.grpc:grpc-netty-shaded:1.16.1

Test: m grpc-netty-shaded-1.16.1-jar
Bug: 148404241
Change-Id: If1784a11df69b61bef0a79476af2231261a9d926
Merged-In: If1784a11df69b61bef0a79476af2231261a9d926