Update Bazel for Windows to version 3.1.0
This change updates the version of Bazel for Windows from a custom
patched version of 1.1.0 to a custom patched version of 3.1.0.

The binary has three cherry-picked commits applied, all of which were in
the original custom patched version of 1.1.0:

* Add --windows_enable_symlinks startup option to support native
symlinks.
  https://github.com/bazelbuild/bazel/commit/cffd417a4c4ec4d082f815f57a771ad1010972dd

* Remove stack size hard limit for WAITER_POOL in WindowsSubprocess.java
  https://github.com/bazelbuild/bazel/commit/08a15e795107ec76d574579223485a3efb9da274

* Check Windows registry if unprivileged symlink creation is supported.
  https://github.com/bazelbuild/bazel/commit/a6fccbb5f28972475f57821ac67db7dd993b2174

The specific patches on top of 3.1.0 for these commits are detailed in
the .patch files in this change.

Bug: 160150358
Test: Manual
Change-Id: I84064f46307b8135511e4123640fde93161359c7