Update TPD binaries
There is a bit of a chicken an egg problem when adding new flags
the PSQ fails, because it is unable to launch TPD with new flag.
This mean that the binaries cannot be pulled from the build
server after the fact. Instead a custom build was queued to
generate these binaries. The version file points to the build
for each OS.

Test: manual
Bug: 169148096
Change-Id: I1402c12a96e67427549f45ee88e359176ab17355