Upgrade to Bazel 4.0.0
We cannot directly upgrade to the official 4.0.0 release because of
https://github.com/bazelbuild/bazel/issues/12972. This only affects
platforms where we use dynamic execution strategies (linux, windows).
The patch file added in this change contains a fix for this issue, and
has been applied on top of custom built bazel 4.0.0 binaries.

Custom bazel binaries were built using:

 bazel build //src:bazel -c opt --stamp --embed_label 4.0.0-b170227190

Windows binary built on a windows-server-2019 image.

Linux binary built on a debian-10 image.

Bug: 160150358
Test: Ran multiple forrest/CI builds using this change
Change-Id: I7e11c0504d377fc6d50b1201c41efaf379731db5