Updating sqlite-inspection.jar
This version contains the following commits:
- http://aosp/Ic0ca12b7381e90c961d0106de979843dc745a9c1
- http://aosp/I24c57f4716f0beddb97e97f72e832bfbfad468ee
- http://aosp/I16129eb28425183bc2f1f016c89ac6a96668574f
- http://aosp/Ifb6affa3e3cb7951e88704d7518326c346fb1d2d

Cherry-pick justification:

This fixes an issue found in stress testing - in scenarios where the
response would be large (e.g. 200 records of 0.5 MB each) we were
observing OutOfMemoryErrors (OOME). These are not currently surfaced to
the user (result of using vanilla Futures).

An alternative to the fix (hardcoding row count limit in response) not
viable as would have to be set low (e.g. 10 rows, so that we can handle
10 x 2MB = 20MB) which would make the tool unusable for e.g. 100,000
small records.

Bug: 161081452
Test: N/A

Change-Id: I7d7252ac49b31434f27a4bc2507779663d1f98f4