Add prebuilt skia library for T.
This is necessary since during T, bazel support is being added to skia
but does not yet support our use case. The presence of BUILD files
prevents us from injecting our own using the normal bazel external repo
mechanism, so for this release the library had to be build manually.

Test: build should work, existing tests should pass.
Bug: 202336345
Change-Id: I41005b2f868efa6146ad8902bc544cf75d2b69d4