Import kotlin-compiler-embeddable:1.6.10
This CL imports kotlin-compiler-embeddable v1.6.10 (same version of the
current Kotlin compiler used in the platform) so that it can be depended
on by other tools (like ktfmt). See [1] for more information.

Most of the content of this CL was generated using the following
commands:

$ curl https://repo1.maven.org/maven2/org/jetbrains/kotlin/kotlin-compiler-embeddable/1.6.10/kotlin-compiler-embeddable-1.6.10.jar --output /tmp/kotlin-compiler-embeddable-1.6.10.jar
$ mvn org.apache.maven.plugins:maven-install-plugin:2.5.1:install-file \
        -Dfile=/tmp/kotlin-compiler-embeddable-1.6.10.jar \
        -DgroupId=org.jetbrains.kotlin \
        -DartifactId=kotlin-compiler-embeddable \
        -Dversion=1.6.10 \
        -Dpackaging=jar \
        -DgeneratePom=true \
        -DlocalRepositoryPath=repository \
        -DcreateChecksum=true

[1] https://groups.google.com/a/google.com/g/android-new-project-reviewers/c/kNOwTqgj5GQ

Bug: 235461679
Test: Manual
Change-Id: I7953a63bf0bf6eff8eab45e008d9d9f9baed3c79
Merged-In: I7953a63bf0bf6eff8eab45e008d9d9f9baed3c79