Add upgraded compose dependency prebuilts
The compose module needs a higher version of the compose-desktop
and skiko-awt dependencies to address the lack of features compose
v1.1.1 has and the bugs skiko-awt v0.7.16 contains. The versions
needed are v1.3.1 (compose-desktop) and v0.7.55 (skiko-awt). In
order to fulfill these dependency needs, the v1.3.1 and v0.7.55
prebuilts of the needed dependencies have been added.

Bug: 272822954
Test: none
Change-Id: I8d579e112f58bbbcb980b1bd23ca1b5c92582474