Update tracing-perfetto-handshake
1. Deleted the checked-in artifact.
2. Re-ran maven-fetch.sh

This means the files are downloaded from a Maven repository,
rather than from a custom ab/ build. This requires deleting
the SBOM file as a result. It will be re-generated based on
the origin.json file.

I don't think we need the sources.jar for studio, so I
removed it from prebuilts.

Bug: 298440275
Test: n/a
Change-Id: I7f31d8036c6ca42caac05ea6309e6d993f6010de