[RESTRICT AUTOMERGE]: Add a "truth" wrapper around "truth-prebuilt"
STS regularly builds tests in sc-dev that get automerged to main.
The automerged tests fail to build on udc-dev because the lib was renamed.

Note: we can't use Merged-In: Iab9b170b08c1e5c7390619e4ec9ec0183a232441
because it's in a different project. Resorting to RESTRICT AUTOMERGE.

Every branch between sc-dev and udc-dev is safe to merge this change.

Bug: 313938773
Test: compile with "truth" instead of "truth-prebuilt"
Change-Id: I2b9ad4c909006a90d8f387c788d18d6acec28231