Update lint-psi to Kotlin 2.0.0-Beta2 and IJ 233
Patches for the old versions are safe to be deleted.
Only one additional change is to enable K2 .kts conversion.

To build 2.0 artifacts, though, we need to adjust language version since
it's 2.0 by default while all other infra is not ready yet.
To reduce the metadata version gap, KGP version is updated too.

Bug: 314713392
Change-Id: Iaaf0c82e1f3e8525508739b5e3f4377a388946d7