Update intellij-rust plugin for IJ platform 232
Although Android Studio is now on IJ 233, this is the most up-to-date
version of the (now technically deprecated) intellij-rust plugin.
However, it's denoted as compatible with up to IJ 233, so there
should not be any issues with this port.

The artifact changes in this CL were generated by running build.py,
which has been modified to reflect the new version.

Bug: 315326466
Test: N/A
Change-Id: Ib4a82e127d6097e46727f19b2ae541fb6cbbd587