Add commonmark-0.22 to prebuilts list
This library already exists in the m2 repo, but was not available
as a library to the Bazel build. It is used by Jewel, but in
I739438f0af3d8cf050c964c7f516fd25c9d0c2f9, we also depend on it
directly in aiplugin/core (we used to depend on 0.20.0, but it is
not binary compatible, and we need to align on the newer one).

Note: this is just the result of running iml_to_build

Bug: none
Test: nothing to test
Change-Id: I7d1902bc05b3bc2eaba082486b4b528fb2fae36d