Update BUILD rule for split arm/x86 TPD binary
This change has no functional impact. It's only a file layout change to
prepare for the future where we will have two trace_processor_daemon
binaries: one for arm64, and another for x86_64.

In this change, the file under the darwin-arm64 directory is identical
to the one under the darwin-x86_64 directory, i.e., it is still x86_64.

Bug: 278924350
Test: n/a
Change-Id: If2832797ca39c546136b8e89aeaa5c09cae64d01