test: Add Wattson reqs to Perfetto config in power CUJs
Add the required events to the Perfetto trace configuration that Wattson
needs to estimate power.

The devfreq trace is needed for accurate power estimation on Pixel 9.
The cpuhp traces are needed to identify the exact timing of CPU being
hotplugged out, which is needed since some power CUJs will excercise
suspend/resume.

Test: ab/I98400010344682914
Bug: 382577589
Change-Id: Icb566c9f89a96dfc957df836e9486178cbd2c1da
Signed-off-by: Samuel Wu <wusamuel@google.com>